requirements.txt
Django==5.0.3
django-debug-toolbar==4.3.0
django-filter==24.2
djangorestframework==3.15.1
psycopg2==2.9.9
drf-spectacular==0.27.0
pre-commit==3.7.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1

